Increase the minimum required Maven version to 3.6.2
Some plugins and dependencies are starting to require higher minimal
Maven version than 3.2.5. Since this is quite an ancient version, it
makes no sense to keep it that old. Also, I believe that many projects
override this default by some higher version already. Let's move this to
3.6.2 now since this is the version used in JBoss EAP/WildFly OpenShift
images. I think it's a nice compromise between being conservative and
moving to latest. This also means minimal required JDK in version 7 [1]
but that should not be any problem either.

[1] https://maven.apache.org/docs/history.html